+++ modules/pcf-exchange/src/main/java/org/eclipse/tractusx/sde/pcfexchange/service/impl/ProxyRequestInterface.java
@@ -98,7 +98,7 @@ public class ProxyRequestInterface {
 	
 	@SneakyThrows
 	public void sendNotificationToConsumer(PCFRequestStatusEnum status, JsonObject calculatedPCFValue,
-			String productId, String bpnNumber, String requestId) {
+			String productId, String bpnNumber, String requestId, String message) {
 
 		// 1 fetch EDC connectors and DTR Assets from EDC connectors
 		List<QueryDataOfferModel> pcfExchangeUrlOffers = edcAssetUrlCacheService.getPCFExchangeUrlFromTwin(bpnNumber);
@@ -112,9 +112,9 @@ public class ProxyRequestInterface {
 			pcfExchangeUrlOffers.parallelStream().forEach(dtOffer -> {
 				
 				if (PCFRequestStatusEnum.SENDING_REJECT_NOTIFICATION.equals(status)) {
-					sendNotification(null, productId, bpnNumber, requestId, dtOffer, status);
+					sendNotification(null, productId, bpnNumber, requestId, dtOffer, status, message);
 				} else {
-					sendNotification(calculatedPCFValue, productId, bpnNumber, requestId, dtOffer, status);
+					sendNotification(calculatedPCFValue, productId, bpnNumber, requestId, dtOffer, status, message);
 				}
 				
 			});
@@ -123,11 +123,9 @@ public class ProxyRequestInterface {
 
 	@SneakyThrows
 	private void sendNotification(JsonObject calculatedPCFValue, String productId, String bpnNumber, String requestId,
-			QueryDataOfferModel dtOffer, PCFRequestStatusEnum status) {
+			QueryDataOfferModel dtOffer, PCFRequestStatusEnum status, String message) {
 		String sendNotificationStatus = "";
 		try {
-			String message = status.name();
-
 			EDRCachedByIdResponse edrToken = edcAssetUrlCacheService.verifyAndGetToken(bpnNumber, dtOffer);
 
 			if (edrToken != null) {