+++ modules/sde-core/src/main/java/org/eclipse/tractusx/sde/configuration/role/CustomPermissionEvaluator.java
@@ -1,6 +1,6 @@
 /********************************************************************************
- * Copyright (c) 2022, 2023 T-Systems International GmbH
- * Copyright (c) 2022, 2023 Contributors to the Eclipse Foundation
+ * Copyright (c) 2022,2024 T-Systems International GmbH
+ * Copyright (c) 2022,2024 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -24,6 +24,7 @@ import java.io.Serializable;
 import java.util.List;
 
 import org.eclipse.tractusx.sde.core.service.RoleManagementService;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.access.PermissionEvaluator;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
@@ -59,6 +60,9 @@ public class CustomPermissionEvaluator implements PermissionEvaluator {
 
 		List<String> list = auth.getAuthorities().stream().map(GrantedAuthority::getAuthority).toList();
 
+		if(list.isEmpty())
+			throw new AccessDeniedException("No access for configured resources");
+		
 		return !roleManagementService.findAll(list, List.of(permissionLs)).isEmpty();
 	}
 }