+++ modules/sde-external-services/edc/src/main/java/org/eclipse/tractusx/sde/edc/constants/EDCAssetConstant.java
@@ -33,14 +33,21 @@ public final class EDCAssetConstant {
 	public static final String ASSET_PROP_FILENAME = "fileName";
 	public static final String ASSET_PROP_TYPE = "type";
 	public static final String ASSET_PROP_POLICYID = "policy-id";
-	public static final String ASSET_PROP_TWIN_VERSION = "1.0.0";
+	public static final String ASSET_PROP_COMMON_VERSION = "1.0.0";
+	public static final String ASSET_PROP_DCAT_VERSION="1.0.0";
 
 	public static final String ASSET_PROP_TYPE_VALUE = "data.core.digitalTwin.submodel";
-	public static final String DATA_CORE_DIGITAL_TWIN_REGISTRY_TYPE = "data.core.digitalTwinRegistry";
-	public static final String DATA_CORE_PCF_EXCHANGE_ENPOINT_TYPE = "data.pcf.exchangeEndpoint";
+	public static final String DATA_CORE_DIGITAL_TWIN_REGISTRY_TYPE = "DigitalTwinRegistry";
+	public static final String DATA_CORE_PCF_EXCHANGE_ENPOINT_TYPE = "PcfExchange";
 
 	public static final String CX_TAXO_PREFIX = "cx-taxo:";
 	public static final String TX_AUTH_PREFIX = "tx-auth:";
+	public static final String CX_POLICY_PREFIX = "cx-policy:";
+	
+	public static final String PCF_FRAMEWORK_AGREEMENT_LEFT_OPERAND = CX_POLICY_PREFIX+"FrameworkAgreement";
+	public static final String PCF_FRAMEWORK_AGREEMENT_RIGHT_OPERAND = "PCF:1.0";
+	public static final String MEMBERSHIP_LEFT_OPERAND = CX_POLICY_PREFIX+"Membership";
+	public static final String ACTIVE_VALUE = "active";
 
 	public static final String DCT_TYPE = "dct:type";
 	public static final String AAS_SEMANTICS_SEMANTIC_ID = "aas-semantics:semanticId";